포인트, 리더보드

+++ src/main/java/com/pgh/album_back/dto/ReviewResponseDTO.java
@@ -10,5 +10,5 @@ public class ReviewResponseDTO {
     private String nickname;
     private String title;
     private String content;
-    private Short rating;
+    private Double rating;
 }
